refactor JSON encoder, move rate-limiting/user-locking to mixins
