+++ authnzerver/frontend/basehandler.py
@@ -28,23 +28,10 @@ from secrets import token_urlsafe
 ## SETUP JSON SERIALIZER ##
 ###########################
 
-class FrontendEncoder(json.JSONEncoder):
-
-    def default(self, obj):
-
-        if isinstance(obj, datetime):
-            return obj.isoformat()
-        elif isinstance(obj, bytes):
-            return obj.decode()
-        elif isinstance(obj, complex):
-            return (obj.real, obj.imag)
-        else:
-            return json.JSONEncoder.default(self, obj)
-
-
 # this replaces the default encoder and makes it so Tornado will do the right
 # thing when it converts dicts to JSON when a
 # tornado.web.RequestHandler.write(dict) is called.
+from .jsonencoder import FrontendEncoder
 json._default_encoder = FrontendEncoder()
