+++ authnzerver/handlers.py
@@ -21,28 +21,12 @@ LOGGER = logging.getLogger(__name__)
 #############
 
 import json
-from datetime import datetime, timedelta
-import asyncio
 from functools import partial
-
-
-class FrontendEncoder(json.JSONEncoder):
-
-    def default(self, obj):
-
-        if isinstance(obj, datetime):
-            return obj.isoformat()
-        elif isinstance(obj, bytes):
-            return obj.decode()
-        elif isinstance(obj, complex):
-            return (obj.real, obj.imag)
-        else:
-            return json.JSONEncoder.default(self, obj)
-
 
 # this replaces the default encoder and makes it so Tornado will do the right
 # thing when it converts dicts to JSON when a
 # tornado.web.RequestHandler.write(dict) is called.
+from .jsonencoder import FrontendEncoder
 json._default_encoder = FrontendEncoder()
 
 import tornado.web
@@ -51,6 +35,7 @@ import tornado.ioloop
 from . import actions
 from .messaging import encrypt_message, decrypt_message
 from .permissions import pii_hash
+from .ratelimit import RateLimitMixin, UserLockMixin
 
 
 #########################
@@ -121,7 +106,9 @@ request_functions = {
 ## MAIN AUTH HANDLER ##
 #######################
 
-class AuthHandler(tornado.web.RequestHandler):
+class AuthHandler(tornado.web.RequestHandler,
+                  RateLimitMixin,
+                  UserLockMixin):
     '''
     This handles the actual auth requests.
 
@@ -154,194 +141,6 @@ class AuthHandler(tornado.web.RequestHandler):
 
         self.allowed_hosts_regex = config.allowed_hosts_regex
         self.ratelimits = config.ratelimits
-
-    def ratelimit_request(self, reqid, request_type, request_body):
-        """
-        This rate-limits the request based on the request type and the
-        set ratelimits passed in the config object.
-
-        """
-
-        # increment the global request each time
-        all_reqcount = self.cacheobj.increment(
-            "all_request_count",
-        )
-
-        # check the global request rate
-        if all_reqcount > self.ratelimits['burst']:
-
-            all_reqrate, all_reqcount, all_req_t0, all_req_tnow = (
-                self.cacheobj.getrate(
-                    "all_request_count",
-                    60.0,
-                )
-            )
-
-            if all_reqrate > self.ratelimits['all']:
-                LOGGER.error(
-                    "[%s] request '%s' is being rate-limited. "
-                    "Cache token: 'all_request_count', count: %s. "
-                    "Rate: %.3f > limit specified for '%s': %s"
-                    % (reqid,
-                       request_type,
-                       all_reqcount,
-                       all_reqrate,
-                       'all',
-                       self.ratelimits['all'])
-                )
-                raise tornado.web.HTTPError(status_code=429)
-
-        # check the user- prefixed request
-        if request_type.startswith("user-"):
-
-            user_cache_token = (
-                request_body.get("email", None) or
-                request_body.get("user_id", None) or
-                request_body.get("session_token", None) or
-                request_body.get("ip_address", None) or
-                request_body.get("email_address", None)
-            )
-
-            # drop all requests that try to get around rate-limiting
-            if not user_cache_token:
-                LOGGER.error(
-                    "[%s] request: '%s' is missing a payload value "
-                    "needed to calculate rate, dropping this request."
-                    % (reqid, request_type)
-                )
-                raise tornado.web.HTTPError(status_code=400)
-
-            user_cache_key = f"user-request-{user_cache_token}"
-
-            user_reqcount = self.cacheobj.increment(
-                user_cache_key,
-            )
-
-            if user_reqcount > self.ratelimits["user"]:
-
-                user_reqrate, user_reqcount, user_req_t0, user_req_tnow = (
-                    self.cacheobj.getrate(
-                        user_cache_key,
-                        60.0,
-                    )
-                )
-
-                if user_reqrate > self.ratelimits['user']:
-                    LOGGER.error(
-                        "[%s] request '%s' is being rate-limited. "
-                        "Cache token: '%s', count: %s. "
-                        "Rate: %.3f > limit specified for '%s': %s"
-                        % (reqid,
-                           request_type,
-                           pii_hash(user_cache_key, self.pii_salt),
-                           user_reqcount,
-                           user_reqrate,
-                           'user',
-                           self.ratelimits['user'])
-                    )
-                    raise tornado.web.HTTPError(status_code=429)
-
-        # check the session- prefixed request
-        if request_type.startswith("session-"):
-
-            session_cache_token = (
-                request_body.get("user_id", None) or
-                request_body.get("session_token", None) or
-                request_body.get("ip_address", None)
-            )
-            if not session_cache_token:
-                LOGGER.error(
-                    "[%s] request: '%s' is missing a payload value "
-                    "needed to calculate rate, dropping this request."
-                    % (reqid, request_type)
-                )
-                raise tornado.web.HTTPError(status_code=400)
-
-            session_cache_key = f"session-request-{session_cache_token}"
-
-            session_reqcount = self.cacheobj.increment(
-                session_cache_key,
-            )
-
-            if session_reqcount > self.ratelimits["session"]:
-
-                (session_reqrate,
-                 session_reqcount,
-                 session_req_t0,
-                 session_req_tnow) = (
-                    self.cacheobj.getrate(
-                        session_cache_key,
-                        60.0,
-                    )
-                )
-
-                if session_reqrate > self.ratelimits['session']:
-                    LOGGER.error(
-                        "[%s] request '%s' is being rate-limited. "
-                        "Cache token: '%s', count: %s. "
-                        "Rate: %.3f > limit specified for '%s': %s"
-                        % (reqid,
-                           request_type,
-                           pii_hash(session_cache_key, self.pii_salt),
-                           session_reqcount,
-                           session_reqrate,
-                           'session',
-                           self.ratelimits['session'])
-                    )
-                    raise tornado.web.HTTPError(status_code=429)
-
-        # check the apikey- prefixed request
-        if request_type.startswith("apikey-"):
-
-            # handle API key issuance
-            apikey_cache_token = (
-                request_body.get("ip_address", None)
-            )
-            # handle all other API key actions
-            if not apikey_cache_token and request_body.get("apikey_dict"):
-                apikey_cache_token = request_body["apikey_dict"].get("uid",
-                                                                     None)
-
-            if not apikey_cache_token:
-                LOGGER.error(
-                    "[%s] request: '%s' is missing a payload value "
-                    "needed to calculate rate, dropping this request."
-                    % (reqid, request_type)
-                )
-                raise tornado.web.HTTPError(status_code=400)
-
-            apikey_cache_key = f"apikey-request-{apikey_cache_token}"
-
-            apikey_reqcount = self.cacheobj.increment(
-                apikey_cache_key,
-            )
-
-            if apikey_reqcount > self.ratelimits["apikey"]:
-
-                (apikey_reqrate,
-                 apikey_reqcount,
-                 apikey_req_t0,
-                 apikey_req_tnow) = (
-                    self.cacheobj.getrate(
-                        apikey_cache_key,
-                        60.0,
-                    )
-                )
-
-                if apikey_reqrate > self.ratelimits['apikey']:
-                    LOGGER.error(
-                        "[%s] request '%s' is being rate-limited. "
-                        "Cache token: '%s', count: %s. "
-                        "Rate: %.3f > limit specified for '%s': %s"
-                        % (reqid,
-                           request_type,
-                           pii_hash(apikey_cache_key, self.pii_salt),
-                           apikey_reqcount,
-                           apikey_reqrate,
-                           'apikey',
-                           self.ratelimits['apikey'])
-                    )
-                    raise tornado.web.HTTPError(status_code=429)
 
     async def send_response(self, response, reqid):
         """
@@ -517,156 +316,3 @@ class AuthHandler(tornado.web.RequestHandler):
 
             LOGGER.exception('Failed to understand request.')
             raise tornado.web.HTTPError(status_code=400)
-
-    async def handle_failed_logins(self, payload):
-        '''
-        This handles failed logins.
-
-        - Adds increasing wait times to successive logins if they keep failing.
-        - If the number of failed logins exceeds 10, the account is locked for
-          one hour, and an unlock action is scheduled on the ioloop.
-
-        '''
-
-        # increment the failure counter and return it
-        if (payload['body']['email'] in self.failed_passchecks):
-            self.failed_passchecks[payload['body']['email']] += 1
-        else:
-            self.failed_passchecks[payload['body']['email']] = 1
-
-        failed_pass_count = self.failed_passchecks[
-            payload['body']['email']
-        ]
-
-        if 0 < failed_pass_count <= self.config.userlocktries:
-            # asyncio.sleep for an exponentially increasing period of time
-            # until 40.0 seconds ~= 10 tries
-            wait_time = 1.5**(failed_pass_count - 1.0)
-            if wait_time > 40.0:
-                wait_time = 40.0
-            await asyncio.sleep(wait_time)
-            return ('wait', failed_pass_count, wait_time)
-
-        elif failed_pass_count > self.config.userlocktries:
-            return ('locked', failed_pass_count, 0.0)
-
-        else:
-            return ('ok', failed_pass_count, 0.0)
-
-    async def lockuser_repeated_login_failures(self,
-                                               payload,
-                                               unlock_after_seconds=3600):
-        '''
-        This locks the user account. Also schedules an unlock action for later.
-
-        '''
-
-        # look up the user ID using the email address
-        loop = tornado.ioloop.IOLoop.current()
-
-        backend_func = partial(
-            actions.get_user_by_email,
-            {'email':payload['body']['email'],
-             'reqid':payload['body']['reqid'],
-             'pii_salt':payload['body']['pii_salt']},
-            config=self.config
-        )
-
-        user_info = await loop.run_in_executor(
-            self.executor,
-            backend_func
-        )
-
-        if not user_info['success']:
-
-            LOGGER.error(
-                "Could not look up the user ID for email: %s to lock "
-                "their account after repeated failed login attempts." %
-                pii_hash(payload['body']['email'], payload['body']['pii_salt'])
-            )
-
-        else:
-
-            # attempt to lock the user using actions.internal_toggle_user_lock
-
-            backend_func = partial(
-                actions.internal_toggle_user_lock,
-                {'target_userid':user_info['user_info']['user_id'],
-                 'action':'lock',
-                 'reqid':payload['body']['reqid'],
-                 'pii_salt':payload['body']['pii_salt']},
-                config=self.config
-            )
-
-            locked_info = await loop.run_in_executor(
-                self.executor,
-                backend_func
-            )
-
-            if locked_info['success']:
-
-                unlock_after_dt = (datetime.utcnow() +
-                                   timedelta(seconds=unlock_after_seconds))
-
-                # schedule the unlock
-                loop.call_later(
-                    unlock_after_seconds,
-                    self.scheduled_user_unlock,
-                    user_info['user_info']['user_id'],
-                    payload['body']['reqid'],
-                    payload['body']['pii_salt']
-                )
-
-                LOGGER.warning(
-                    "Locked the account for user ID: %s, "
-                    "email: %s after repeated "
-                    "failed login attempts. "
-                    "Unlock scheduled for: %sZ" %
-                    (pii_hash(user_info['user_info']['user_id'],
-                              payload['body']['pii_salt']),
-                     pii_hash(payload['body']['email'],
-                              payload['body']['pii_salt']),
-                     unlock_after_dt)
-                )
-
-        # we'll return a failure here as the response no matter what happens
-        # above to deny the login
-        return {
-            "success": False,
-            "user_id":None,
-            "messages":[
-                "Your user account has been locked "
-                "after repeated login failures. "
-                "Try again in an hour or "
-                "contact the server admins."
-            ]
-        }
-
-    async def scheduled_user_unlock(self, user_id, reqid, pii_salt):
-        '''
-        This function is scheduled on the ioloop to unlock the specified user.
-
-        '''
-
-        LOGGER.warning(
-            "[%s] Unlocked the account for user ID: %s after "
-            "login-failure timeout expired." %
-            (reqid, pii_hash(user_id, pii_salt))
-        )
-
-        loop = tornado.ioloop.IOLoop.current()
-
-        backend_func = partial(
-            actions.internal_toggle_user_lock,
-            {'target_userid':user_id,
-             'action':'unlock',
-             'reqid':reqid,
-             'pii_salt':pii_salt},
-            config=self.config
-        )
-
-        locked_info = await loop.run_in_executor(
-            self.executor,
-            backend_func
-        )
-        return locked_info
